app/azure_functions/ri_collector/requirements.txt
azure-functions
azure-storage-blob
azure-identity
psycopg2-binary
requests

